attempting to refactor
