+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/metadata/ProjectSampleMetadataController.java
@@ -1,18 +1,11 @@
 package ca.corefacility.bioinformatics.irida.ria.web.projects.metadata;
 
 import java.io.*;
-import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.Map.Entry;
 
 import javax.servlet.http.HttpSession;
 
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVRecord;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import org.apache.poi.ss.usermodel.*;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,11 +22,11 @@ import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.sample.metadata.MetadataEntry;
 import ca.corefacility.bioinformatics.irida.ria.utilities.SampleMetadataStorage;
 import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectControllerUtils;
+import ca.corefacility.bioinformatics.irida.ria.web.services.UIMetadataImportService;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
 
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Files;
@@ -51,16 +44,18 @@ public class ProjectSampleMetadataController {
 	private final SampleService sampleService;
 	private final MetadataTemplateService metadataTemplateService;
 	private final ProjectControllerUtils projectControllerUtils;
+	private final UIMetadataImportService metadataImportService;
 
 	@Autowired
 	public ProjectSampleMetadataController(MessageSource messageSource, ProjectService projectService,
 			SampleService sampleService, MetadataTemplateService metadataTemplateService,
-			ProjectControllerUtils projectControllerUtils) {
+			ProjectControllerUtils projectControllerUtils, UIMetadataImportService metadataImportService) {
 		this.messageSource = messageSource;
 		this.projectService = projectService;
 		this.sampleService = sampleService;
 		this.metadataTemplateService = metadataTemplateService;
 		this.projectControllerUtils = projectControllerUtils;
+		this.metadataImportService = metadataImportService;
 	}
 
 	/**
@@ -80,102 +75,26 @@ public class ProjectSampleMetadataController {
 		// We want to return a list of the table headers back to the UI.
 		SampleMetadataStorage storage = new SampleMetadataStorage();
 		try {
-			// Need an input stream
 			String filename = file.getOriginalFilename();
 			byte[] byteArr = file.getBytes();
-			InputStream fis = new ByteArrayInputStream(byteArr);
+			InputStream inputStream = new ByteArrayInputStream(byteArr);
 
-			Workbook workbook = null;
 			String extension = Files.getFileExtension(filename);
 
 			// Check the file type
 			switch (extension) {
 			case "csv":
-				CSVParser parser = CSVParser.parse(fis, StandardCharsets.UTF_8,
-						CSVFormat.RFC4180.withFirstRecordAsHeader()
-								.withTrim()
-								.withIgnoreEmptyLines());
-				List<Map<String, String>> rows = new ArrayList<>();
-
-				// save headers
-				Map<String, Integer> headers_set = parser.getHeaderMap();
-				List<String> headers_list = new ArrayList<>(headers_set.keySet());
-				storage.saveHeaders(headers_list);
-
-				// save data
-				for (CSVRecord row : parser) {
-					System.out.println(row);
-					Map<String, String> rowMap = new HashMap<>();
-					for (String key : row.toMap()
-							.keySet()) {
-						String value = row.toMap()
-								.get(key);
-						rowMap.put(key, value);
-					}
-					rows.add(rowMap);
-				}
-
-				storage.saveRows(rows);
-				parser.close();
-				fis.close();
-				session.setAttribute("pm-" + projectId, storage);
-				return storage;
-			// Check the type of workbook
-			case "xlsx":
-				workbook = new XSSFWorkbook(fis);
+				storage = metadataImportService.parseCSV(inputStream);
 				break;
+			case "xlsx":
 			case "xls":
-				workbook = new HSSFWorkbook(fis);
+				storage = metadataImportService.parseExcel(inputStream, extension);
 				break;
 			default:
 				// Should never reach here as the uploader limits to .csv, .xlsx and .xlx files.
 				throw new MetadataImportFileTypeNotSupportedError(extension);
 			}
-
-			// Only look at the first sheet in the workbook as this should be the file we want.
-			Sheet sheet = workbook.getSheetAt(0);
-			Iterator<Row> rowIterator = sheet.iterator();
-
-			List<String> headers = getWorkbookHeaders(rowIterator.next());
-			storage.saveHeaders(headers);
-
-			// Get the metadata out of the table.
-			List<Map<String, String>> rows = new ArrayList<>();
-			while (rowIterator.hasNext()) {
-				Map<String, String> rowMap = new HashMap<>();
-				Row row = rowIterator.next();
-				Iterator<Cell> cellIterator = row.cellIterator();
-				while (cellIterator.hasNext()) {
-					Cell cell = cellIterator.next();
-
-					int columnIndex = cell.getColumnIndex();
-					if (columnIndex < headers.size()) {
-						String header = headers.get(columnIndex);
-
-						if (!Strings.isNullOrEmpty(header)) {
-							// Need to ignore empty headers.
-							if (cell.getCellTypeEnum()
-									.equals(CellType.NUMERIC)) {
-								/*
-								This is a special handler for number cells.  It was requested that numbers
-								keep their formatting from their excel files.  E.g. 2.222222 with formatting
-								for 2 decimal places will be saved as 2.22.
-								 */
-								DataFormatter formatter = new DataFormatter();
-								String value = formatter.formatCellValue(cell);
-								rowMap.put(header, value);
-							} else {
-								cell.setCellType(CellType.STRING);
-								rowMap.put(header, cell.getStringCellValue());
-							}
-						}
-					}
-				}
-				rows.add(rowMap);
-			}
-			storage.saveRows(rows);
-
-			fis.close();
+			inputStream.close();
 		} catch (FileNotFoundException e) {
 			logger.debug("No file found for uploading an excel file of metadata.");
 		} catch (IOException e) {
@@ -184,37 +103,6 @@ public class ProjectSampleMetadataController {
 
 		session.setAttribute("pm-" + projectId, storage);
 		return storage;
-	}
-
-	/**
-	 * Extract the headers from an excel file.
-	 *
-	 * @param row {@link Row} First row from the excel file.
-	 * @return {@link List} of {@link String} header values.
-	 */
-	private List<String> getWorkbookHeaders(Row row) {
-		// We want to return a list of the table headers back to the UI.
-		List<String> headers = new ArrayList<>();
-
-		// Get the column headers
-		Iterator<Cell> headerIterator = row.cellIterator();
-		while (headerIterator.hasNext()) {
-			Cell headerCell = headerIterator.next();
-			CellType cellType = headerCell.getCellTypeEnum();
-
-			String headerValue;
-			if (cellType.equals(CellType.STRING)) {
-				headerValue = headerCell.getStringCellValue()
-						.trim();
-			} else {
-				headerValue = String.valueOf(headerCell.getNumericCellValue())
-						.trim();
-			}
-
-			// Leave empty headers for now, we will remove those columns later.
-			headers.add(headerValue);
-		}
-		return headers;
 	}
 
 	/**
